This commit makes a bunch of changes.  I made the package and tests better organized.  I also dramatically changed docket_query.
Now, docket_query, the heart of this little program, works like
MapReduce.  User creates a function to apply to every docket,
then uses a class in docket_query to apply that function to
each docket and get the results as hashes.

The next step will be to add a helper function that will turn a
list of hashes into csv rows and save the resulting csv table
to a file.